Extracted enum to ID mapping to its own helper class

+++ src/test/java/com/gagror/data/account/AccountTypeConverterUnitTest.java
@@ -23,7 +23,7 @@ public class AccountTypeConverterUnitTest {
 
 	@Test
 	public void convertToDatabaseColumn() {
-		assertEquals(accountType.getId(), instance.convertToDatabaseColumn(accountType).intValue());
+		assertEquals(accountType.getId(), instance.convertToDatabaseColumn(accountType));
 	}
 
 	@Test
